Added methods to compute EIP-1559 gas pool and legacy gas pool. Few minor changes: - Renaming. - Bug fixing.
