+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/BlockUtils.java
@@ -48,6 +48,7 @@ public class BlockUtils {
         block.getGasUsed().longValue(),
         block.getTimestamp().longValue(),
         Bytes.fromHexString(block.getExtraData()),
+        0,
         mixHash,
         block.getNonce().longValue(),
         blockHeaderFunctions);
