+++ ethereum/core/src/test-support/java/org/hyperledger/besu/ethereum/core/NonBesuBlockHeader.java
@@ -110,4 +110,9 @@ public class NonBesuBlockHeader implements BlockHeader {
   public Hash getBlockHash() {
     return blockHash;
   }
+
+  @Override
+  public long getBaseFee() {
+    return 0;
+  }
 }
